plugins: split workspace-wall into a separate helper library
