+++ plugins/single_plugins/meson.build
@@ -8,10 +8,16 @@ plugins = [
 all_include_dirs = [wayfire_api_inc, wayfire_conf_inc, plugins_common_inc, vswitch_inc, wobbly_inc, grid_inc]
 all_deps = [wlroots, pixman, wfconfig, wftouch, cairo, pango, pangocairo, json, plugin_pch_dep]
 
+extra_libs = {
+    'expo': [workspace_wall],
+    'vswipe': [workspace_wall],
+}
+
 foreach plugin : plugins
   shared_module(plugin, plugin + '.cpp',
       include_directories: all_include_dirs,
       dependencies: all_deps,
+      link_with: extra_libs.get(plugin, []),
       install: true,
       install_dir: conf_data.get('PLUGIN_PATH'))
 endforeach
