+++ plugins/vswitch/meson.build
@@ -4,6 +4,7 @@ vswitch = shared_module('vswitch',
     ['vswitch.cpp'],
     include_directories: [wayfire_api_inc, wayfire_conf_inc, plugins_common_inc, vswitch_inc, ipc_include_dirs],
     dependencies: [wlroots, pixman, wfconfig, json, plugin_pch_dep],
+    link_with: [workspace_wall],
     install: true,
     install_dir: join_paths(get_option('libdir'), 'wayfire'))
